Rename RequestPasswordResetView
