+++ packages/client/src/app/auth/views/RequestPasswordResetView.vue
