+++ packages/client/src/app/auth/views/index.js
@@ -1,5 +1,5 @@
 import LoginView from './LoginView.vue';
 import RegisterView from './RegisterView.vue';
-import ResetPasswordView from './ResetPasswordView.vue';
+import RequestPasswordResetView from './RequestPasswordResetView.vue';
 
-export { LoginView, RegisterView, ResetPasswordView };
+export { LoginView, RegisterView, RequestPasswordResetView };
